+++ tests/integration_tests/functional/test_mmds.py
@@ -824,8 +824,8 @@ def test_mmds_snapshot(bin_cloner_path,  version):
     # Create a snapshot with current build and restore with each FC binary
     # artifact.
     firecracker_artifacts = artifacts.firecrackers(
-        # v1.0.0 breaks snapshot compatibility with older versions.
-        min_version="1.0.0",
+        # v1.1.0 breaks snapshot compatibility with older versions.
+        min_version="1.1.0",
         max_version=get_firecracker_version_from_toml())
     for firecracker in firecracker_artifacts:
         vm_instance = vm_builder.build_vm_nano(